Complete Extension - B-Reminders
Added in a new extension feature for reminders that allow users to look
at not just the reminders for tasks  happening today, but also for those
happening up to tomorrow, up to next week, or even all the reminders all
at once. Also improved on some of the Coding Standard and Code Quality Conventions.